app/pyproject.toml
[tool.poetry]
name = "image-rec"
version = "0.1.0"
description = ""
authors = ["Clyde Tedrick <clyde.tedrick@zagaopensource.com>"]

[tool.poetry.dependencies]
python = "^3.8"
torch = "*"
torchvision = "*"
pillow = "*"
opencv-python = "*"

[tool.poetry.dev-dependencies]
pytest = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
